+++ pkg/project/version_bundle_aws.go
@@ -8,9 +8,11 @@ var versionBundleAWS = versionbundle.Bundle{
 	Changelogs: []versionbundle.Changelog{
 		{
 			Component:   Name(),
-			Description: "TODO",
+			Description: "Enable NodePort ingress service on KVM.",
 			Kind:        versionbundle.KindChanged,
-			URLs:        []string{},
+			URLs: []string{
+				"https://github.com/giantswarm/cluster-operator/pull/1162",
+			},
 		},
 	},
 	Components: []versionbundle.Component{},